requirements.txt
Pillow==9.0.1
setuptools==58.0.4
streamlit==1.5.1
